+++ packages/tempo/test/core/accessors.test.ts
@@ -18,11 +18,13 @@ describe(`${label}`, () => {
     expect(tempo.dd).toBe(date.getDate())
   })
 
-  test(`${label} get instance locale via loc getters`, () => {
+  test(`${label} get instance locale and sphere getters`, () => {
     const tDefault = new Tempo('2024-05-20');
-    expect(tDefault.loc).toBeDefined();
+    expect(tDefault.locale).toBeDefined();
+    expect(tDefault.sphere).toBeDefined();
 
-    const tCustom = new Tempo('2024-05-20', { locale: 'fr-FR' });
-    expect(tCustom.loc).toBe('fr-FR');
+    const tCustom = new Tempo('2024-05-20', { locale: 'fr-FR', sphere: 'south' });
+    expect(tCustom.locale).toBe('fr-FR');
+    expect(tCustom.sphere).toBe('south');
   })
 })